+++ react-native-purchases-ui/android/src/main/java/com/revenuecat/purchases/react/ui/RNPaywallsModule.kt
@@ -1,21 +1,15 @@
 package com.revenuecat.purchases.react.ui
 
-import android.app.Activity
-import android.content.Intent
 import android.util.Log
 import androidx.fragment.app.FragmentActivity
-import com.facebook.react.bridge.ActivityEventListener
-import com.facebook.react.bridge.BaseActivityEventListener
 import com.facebook.react.bridge.Promise
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
 import com.facebook.react.bridge.ReactMethod
-import com.revenuecat.purchases.PurchasesErrorCode
 import com.revenuecat.purchases.hybridcommon.ui.PaywallResultListener
 import com.revenuecat.purchases.hybridcommon.ui.PaywallSource
 import com.revenuecat.purchases.hybridcommon.ui.PresentPaywallOptions
 import com.revenuecat.purchases.hybridcommon.ui.presentPaywallFromFragment
-import com.revenuecat.purchases.ui.revenuecatui.customercenter.ShowCustomerCenter
 
 
 internal class RNPaywallsModule(
@@ -24,7 +18,6 @@ internal class RNPaywallsModule(
 
     companion object {
         const val NAME = "RNPaywalls"
-        private const val REQUEST_CODE_CUSTOMER_CENTER = 1001
     }
 
     private val currentActivityFragment: FragmentActivity?
@@ -37,37 +30,6 @@ internal class RNPaywallsModule(
                 }
             }
         }
-
-    private var customerCenterPromise: Promise? = null
-
-    private val activityEventListener: ActivityEventListener =
-        object : BaseActivityEventListener() {
-            override fun onActivityResult(
-                activity: Activity,
-                requestCode: Int,
-                resultCode: Int,
-                intent: Intent?
-            ) {
-                if (requestCode == REQUEST_CODE_CUSTOMER_CENTER) {
-                    if (resultCode == Activity.RESULT_OK) {
-                        Log.d(NAME, "Customer Center closed successfully")
-                        customerCenterPromise?.resolve(null)
-                    } else {
-                        Log.d(NAME, "Customer Center closed with result $resultCode")
-                        customerCenterPromise?.reject(
-                            PurchasesErrorCode.UnknownError.code.toString(),
-                            "Customer Center closed with result code: $resultCode",
-                            null
-                        )
-                    }
-                    customerCenterPromise = null
-                }
-            }
-        }
-
-    init {
-        reactContext.addActivityEventListener(activityEventListener)
-    }
 
     override fun getName(): String {
         return NAME
@@ -107,22 +69,6 @@ internal class RNPaywallsModule(
     }
 
     @ReactMethod
-    fun presentCustomerCenter(
-        promise: Promise
-    ) {
-        currentActivity?.let {
-            presentCustomerCenterFromActivity(it)
-            customerCenterPromise = promise
-        } ?: run {
-            promise.reject(
-                PurchasesErrorCode.UnknownError.code.toString(),
-                "Could not present Customer Center. There's no activity",
-                null
-            )
-        }
-    }
-
-    @ReactMethod
     fun addListener(eventName: String?) {
         // Keep: Required for RN built in Event Emitter Calls.
     }
@@ -159,13 +105,5 @@ internal class RNPaywallsModule(
                 fontFamily = fontFamily
             )
         )
-    }
-
-    private fun presentCustomerCenterFromActivity(
-        activity: Activity
-    ) {
-        val intent = ShowCustomerCenter()
-            .createIntent(activity, Unit)
-        activity.startActivityForResult(intent, REQUEST_CODE_CUSTOMER_CENTER)
     }
 }
