+++ react-native-purchases-ui/android/src/main/java/com/revenuecat/purchases/react/ui/RevenueCatUIPackage.kt
@@ -6,9 +6,9 @@ import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 
 
-class RNPaywallsPackage : ReactPackage {
+class RevenueCatUIPackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-        return listOf(RNPaywallsModule(reactContext))
+        return listOf(RNPaywallsModule(reactContext), RNCustomerCenterModule(reactContext))
     }
 
     override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
